Leave arch mappings in _project_options for now
